Add MIT-0, LGPLv2.1 and APSL-2.0 to licenses
Three new licenses are added in this commit: MIT-0, LGPLv2.1, APSL-2.0.
The short name for these licenses corresponds to their SPDX identifier, except for LGPLV2.1.
The logn name is similar to the closest license, for example LGPLv3 for LGPLv2.1.

The source of these lincenses is [Open Source Initiative](https://opensource.org/licenses/alphabetical).