Added Open Source Initiative as licenses source
Added the link to the [Open Source Initiative](https://opensource.org/licenses/) license list next to choosealicense.com.
Like choosealicense.com, it is added as comment on top of LICENSES declaration.

The information of the licenses was extracted from the alphabetical ordered list, although a more generic link is provided here.