Skeleton bash instaler
A skeleton for the bash command line interface instaler is created.
So far, it checks if you have any internet fetching utility installed
for installation, if you have never installed lsfp before for update and
a tip for typing install if yuo typed some similar command but not
install.
More work will be put on this.