Icons (& other optimizations)
This commit adds icons to lsfp.
It makes use of Nerd Fonts icon aggregation, which makes a bunch of
icons available as unicode characters in a font.
Icons are disabled by default, therefore you need to pass the -i/--icons
flag or set the icons field in a .lsfprc file.
The corresponding option and config help sections were added.

To get icons going, a few modifications to the code were made.
The `file_detection::file_extension_color` function has been changed to
return a tuple, with its first item being the color and the second icon.
The FILE_EXTENSION_COLORS constant has now a third item in its tuples,
which is a string literal with the matching icon escaped unicode.
Other icons (folders, license and generic) have also been added as
constants to `constsants.rs`.
A leading space before the icon and between it and the name has also
been added to improve icons visualization. When running in tree format,
this space is also added between the tree branch and the icon.

Other than the icons, there are some small improvements to the code.
`file_detection::file_extension_styles` has been modified so no mutable
variable is needed and the return type has also been changed, as well as
the name.
The "Config" section of help has also been modified to only show when
the config feature is on, as well as each individual entry inside it.