Themes
This commit implements themes, discussed in issue #56.
It brings the parser created aside in the repository
[HipyCas/theme-parser](https://github.com/HipyCas/theme-parser) and
adapts it to the project.
Themes are passed using the `--theme` flag followed by the name or path.
For more information, read the [README.md of the `themes`
module](https://github.com/The-Noah/lsfp/blob/master/src/themes/README.md)
or for theme file syntax the [formal
specification](https://github.com/The-Noah/lsfp/blob/master/src/themes/SPEC.md).

**Features**

With this addition, three new features where added.
The feature that enables themes is _themes_, that at the same time depend on other features.
The pairs that get parsed in the theme file depend at the same time on which other
features are enabled (_color_ and _icons_).
Other two features added which are only for internal use, in which _themes_ depends, are _home_ (adds
function to get home directory) and _index_flags_ (enables indexing of
`lsfp::core::args::Flags`).

**Named themes**

`lsfp` can load a theme by its name or by path. To load a specific path
make sure that it exists and includes either a dot or a forward or
backward slash.
When loading a theme by name, `lsfp` looks for a file named like the
passed name under `~/.lsfp-themes`.