Chainable single letter flags (#62)
With this commit, the possibility of chaining options/flags represented
by a single letter is brought in. It works as many other CLIs do, and so
passing "-atis" is equivalent to passing individually each on ("-a -t -i
-s").

This is implemented by applying `flat_map` on the args iteratior (after
having applied the `skip` function to skip the first argument, i.e.
program path), which filters for options that start with a dash, and
followed by at least two more letters and no dash. The implementation
may look a bit verbose, but is probably as simple as it gets.