Remove excessive line breaks in config print
This commits removes an extra newline added to the `set <key>=<value>`
message printed when setting configuration via the CLI options. In the
previous commit, the behaviour of the CLI when setting config to not run
the folder scanning and printing, instead just print the configuration
message. This commit closes this updates.